# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clothing e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
  staff: 'staff',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  user: 'user',
  admin: 'admin',
  staff: 'staff',
} as const;

export interface UserUpdate {
  name?: string;
  email?: string;
  role?: UserUpdateRole;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
}

export interface Product {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  category: string;
  /** @nullable */
  categoryId?: number | null;
  images: string[];
  /** @nullable */
  modelUrl?: string | null;
  stock: number;
  sizes: string[];
  colors: string[];
  tags?: string[];
  featured?: boolean;
  createdAt: string;
}

export interface ProductInput {
  title: string;
  description?: string;
  price: number;
  compareAtPrice?: number;
  category: string;
  categoryId?: number;
  images?: string[];
  modelUrl?: string;
  stock: number;
  sizes?: string[];
  colors?: string[];
  tags?: string[];
  featured?: boolean;
}

export interface ProductUpdate {
  title?: string;
  description?: string;
  price?: number;
  compareAtPrice?: number;
  category?: string;
  categoryId?: number;
  images?: string[];
  modelUrl?: string;
  stock?: number;
  sizes?: string[];
  colors?: string[];
  tags?: string[];
  featured?: boolean;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface CartItem {
  id: number;
  productId: number;
  quantity: number;
  /** @nullable */
  size: string | null;
  /** @nullable */
  color: string | null;
  product: Product;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
  size?: string;
  color?: string;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface WishlistItem {
  id: number;
  productId: number;
  product: Product;
  createdAt: string;
}

export interface WishlistInput {
  productId: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  quantity: number;
  price: number;
  /** @nullable */
  size: string | null;
  /** @nullable */
  color: string | null;
  product: Product;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type OrderPaymentStatus = typeof OrderPaymentStatus[keyof typeof OrderPaymentStatus];


export const OrderPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface Address {
  id: number;
  userId: number;
  /** @nullable */
  label?: string | null;
  line1: string;
  /** @nullable */
  line2?: string | null;
  city: string;
  state: string;
  country: string;
  zip: string;
  isDefault?: boolean;
}

export interface Order {
  id: number;
  userId: number;
  items: OrderItem[];
  totalPrice: number;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  address?: Address;
  user?: User;
  createdAt: string;
  updatedAt?: string;
}

export interface OrderInput {
  addressId: number;
  paymentMethod: string;
  couponCode?: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type OrderStatusUpdatePaymentStatus = typeof OrderStatusUpdatePaymentStatus[keyof typeof OrderStatusUpdatePaymentStatus];


export const OrderStatusUpdatePaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
  paymentStatus?: OrderStatusUpdatePaymentStatus;
}

export interface AddressInput {
  label?: string;
  line1: string;
  line2?: string;
  city: string;
  state: string;
  country: string;
  zip: string;
  isDefault?: boolean;
}

export interface RevenueByDay {
  date: string;
  revenue: number;
  orders: number;
}

export interface OrdersByStatus {
  status: string;
  count: number;
}

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  totalUsers: number;
  revenueByDay: RevenueByDay[];
  ordersByStatus: OrdersByStatus[];
}

export type ListProductsParams = {
category?: string;
search?: string;
minPrice?: number;
maxPrice?: number;
size?: string;
color?: string;
page?: number;
limit?: number;
};