# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clothing e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from './address';
import type { OrderItem } from './orderItem';
import type { OrderPaymentStatus } from './orderPaymentStatus';
import type { OrderStatus } from './orderStatus';
import type { User } from './user';

export interface Order {
  id: number;
  userId: number;
  items: OrderItem[];
  totalPrice: number;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  address?: Address;
  user?: User;
  createdAt: Date;
  updatedAt?: Date;
}