# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clothing e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  category: string;
  /** @nullable */
  categoryId?: number | null;
  images: string[];
  /** @nullable */
  modelUrl?: string | null;
  stock: number;
  sizes: string[];
  colors: string[];
  tags?: string[];
  featured?: boolean;
  createdAt: Date;
}